Include whitespace-only text nodes
Also includes whitespace at the beginning of a text node.

Without this, we throw away information needed to create an exact
round-trip copy of the input, as well as information needed to create
a canonicalized version of the XML Document.

Should only affect the .nodes() interface, as well as .contents().
.elements() and XML::Text.string should be unaffected. Note that
XML::Text.Str may be affected.