Fixed parsing of XML with preambles.
Xml did not correctly parse and handle the XML preamble. It was storeing the
version and encoding with the quotes. It was also not working correctly with
preambles that used a single quote rather than a ".

NOTE!!! this modules does NOT handle the standalone parameter of the preamble
(the preamble is also known as the "Declaration")